moved bucky routes
